chore: isolate wasm bindings into a separate file
