+++ packages/algorithms/CMakeLists.txt
@@ -9,8 +9,11 @@ set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 
 add_executable(${PROJECT_NAME}
+  src/add.h
   src/add.cpp
+  src/bbb.h
   src/bbb.cpp
+  src/wasm.cpp
   )
 
 set_property(TARGET ${PROJECT_NAME} PROPERTY C_STANDARD 11)
