+++ packages/algorithms/src/add.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <string>
+
+int add(int x, int y);
+
+std::string concat(const std::string& x, const std::string& y);
+
+void kaboom();
